Added CMD Prompt (Windows) compatability
Originally, this "OS" only worked with linux, since I used "system('clear')" to clear the screen, but that didn't work with windows, so I made a function for clearing the screen that checks if the terminal the py file is opened in a linux like terminal or if it's opened in windows command prompt. Which makes it so that you can now use this OS with command prompt in windows.